Corrected implementation of Timer.
Appears last implementation had inconsistencies and would not work. Tested this implementation on my local machine, appears to work as expected.

More testing may be a good idea b/c the only place this is used (currently) is PID, so we won't be able to tell if it works or not.