Rotation sensors (#36)
* Moving motor and odometry initialization into constructor.

This just seems like a more robust location for these actions to be preformed. This may also allow resistence to bugs where the `initialize` method is not called.

* [Implementation] `Drivetrain::get_position()` now uses rotation sensors.

Odometry code does not need to change, as all reading of the positions is done through the one method that was updated.

We need to make sure the ports I used are correct in the lab.

* Fixed port numbers (I think)

* Moving motor and odometry initialization into constructor.

This just seems like a more robust location for these actions to be preformed. This may also allow resistence to bugs where the `initialize` method is not called.

* [Implementation] `Drivetrain::get_position()` now uses rotation sensors.

Odometry code does not need to change, as all reading of the positions is done through the one method that was updated.

We need to make sure the ports I used are correct in the lab.

* Fixed port numbers (I think)

* Updated timer for intuition and flexibility.

Previously, the timer would need to be stopped before getting elapsed time. Now, elapsed time can be obtained without stopped. `timer::start()` also acts as a reset.

* [Implementation] Enhance odometry calculations with angular velocity thresholding

When angular velocity exceeds some threshold, we use arc-based odometry. When below that threshold, we use linear odometry. Found that this increases accuracy a ton for straight paths.

The threshold is set to 4 degrees/second.

* Reset rotation sensor positions during initialization

* [Doc] Clarify tracking wheel measurement description in Constants.hpp

---------

Co-authored-by: mtdaly <mtdaly@ncsu.edu>