[Debug] Correcting `Command`'s `run()` method.
Previously, the loop counter was only reflecting the number of times a `Command`'s `periodic` was run. This causes issues where `InstantCommand`s execute twice because their `periodic` is never called. Now, the loop count reflects the number of times the `run` method is called, fixing this issue for `InstantCommand`s.